Rename module `core.cal_steps` -> `core.recipe`
